teste

+++ paginas/dash.py
@@ -22,7 +22,7 @@ df['RA'] = df['RA'].astype(int)
 bd = ler_sheets_cache('bd')
 bd['RA'] = bd['RA'].astype(int)
 df_historico = ler_sheets_cache('historico')
-df_historico['RA'] = df_historico['RA'].astype(int)
+#df_historico['RA'] = df_historico['RA'].astype(int)
 
 
 #filtros
